rename package referendumvorlage -> referendumvorlagen
