+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/rest/referendumvorlagen/ReferendumvorlagenDTOMapper.java
@@ -1,0 +1,13 @@
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.rest.referendumvorlagen;
+
+import de.muenchen.oss.wahllokalsystem.basisdatenservice.services.referendumvorlagen.ReferendumvorlagenModel;
+import de.muenchen.oss.wahllokalsystem.basisdatenservice.services.referendumvorlagen.ReferendumvorlagenReferenceModel;
+import org.mapstruct.Mapper;
+
+@Mapper
+public interface ReferendumvorlagenDTOMapper {
+
+    ReferendumvorlagenDTO toDTO(ReferendumvorlagenModel referendumvorlagenModel);
+
+    ReferendumvorlagenReferenceModel toModel(String wahlbezirkID, String wahlID);
+}
