+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/services/referendumvorlagen/ReferendumvorlagenClient.java
@@ -1,4 +1,4 @@
-package de.muenchen.oss.wahllokalsystem.basisdatenservice.services.referendumvorlage;
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.services.referendumvorlagen;
 
 public interface ReferendumvorlagenClient {
